Refactoring. Renamed lorenz to attractor to allow for the definition of other equation systems such as rossler.
