+++ software/contrib/lorenz.py
@@ -1,188 +1,83 @@
-from europi import *
-from time import ticks_diff, ticks_ms
-from math import fabs, floor
-from attractor import attractor
-
 '''
-Lorenz
-author: Sean Bechhofer (github.com/seanbechhofer)
-date: 2022-03-15
-labels: gates, triggers, randomness
+Implementation of a simple Lorenz Attractor, see 
 
-Lorenz is a source of chaotic modulation based on a Lorenz Attractor: https://en.wikipedia.org/wiki/Lorenz_system
+https://en.wikipedia.org/wiki/Lorenz_system
 
-Outputs 1, 2 and 3 are based on the x, y and z values generated by the attractor.
-
-Outputs 4, 5 and 6 are gates based on the values of x, y and z and relationships between them. 
-
-digital_in: Pause motion when HIGH
-analog_in: 
-
-knob_1: Adjust speed
-knob_2: Adjust threshold for triggers
-
-button_1: Decrease output voltage range
-button_2: Increase output voltage range
-
-output_1: x 
-output_2: y
-output_3: z
-output_4: triggers/gates
-output_5: triggers/gates
-output_6: triggers/gates
-
+Default uses well known values of s=10,r=28,b=2.667. 
 '''
 
-# Overclock the Pico for improved performance.
-machine.freq(250_000_000)
-
-# Maximum voltage output. Cranking this up may cause issues with some modules. 
-MAX_OUTPUT = 5
-
-class lorenz:
-    def __init__(self):
-        # create the attractor and initialise.
-        # This will take around 30 seconds.
-        self.a = attractor()
-        self.initialiseMessage()
-        self.a.estimateRanges()
+class attractor:
+    def __init__(self, point=(0.,1.,1.05), params=(10,28,2.667), dt=0.01):
+        self.x = point[0]
+        self.y = point[1]
+        self.z = point[2]
+        self.s = params[0]
+        self.r = params[1]
+        self.b = params[2]
+        self.dt = dt
+        self.x_min = self.x
+        self.y_min = self.y
+        self.z_min = self.z
+        self.x_max = self.x
+        self.y_max = self.y
+        self.z_max = self.z
+        # arbitrary initial range values    
+        self.x_range = 100
+        self.y_range = 100
+        self.z_range = 100
+        
+    # The range of values produced depends on the parameters. If we
+    # know the range, we can then normalise coordinates for use when
+    # generating CV. This method runs through a number of iterations
+    # to estimate ranges.
+    def estimateRanges(self,steps=100000):
+    
+        # Execute a number of steps to get upper and lower bounds. 
+        for i in range(steps):
+            self.step()
             
-        # Initialize variables
-        self.checkpoint = 0
-        # time before update
-        self.period = 100
-        # output range.
-        self.range = MAX_OUTPUT
-        # initial threshold for gates
-        self.threshold = 20
-        # freeze motion
-        self.freeze = False
+            self.x_max = max(self.x, self.x_max)
+            self.y_max = max(self.y, self.y_max)
+            self.z_max = max(self.z, self.z_max)
+            self.x_min = min(self.x, self.x_min)
+            self.y_min = min(self.y, self.y_min)
+            self.z_min = min(self.z, self.z_min)
 
-        # Triggered when button 1 is released
-        # Short press: decrease range
-        # Long press: 
-        @b1.handler_falling
-        def b1Pressed():
-            if ticks_diff(ticks_ms(), b1.last_pressed()) >  300:
-                # long press
-                pass
-            else:
-                # short press
-                self.range -= 1
-                if self.range < 1:
-                    self.range = 1
+        self.x_range = self.x_max-self.x_min
+        self.y_range = self.y_max-self.y_min
+        self.z_range = self.z_max-self.z_min
 
-        # Triggered when button 2 is released.
-        # Short press: increase range
-        # Long press: 
-        @b2.handler_falling
-        def b2Pressed():
-            
-            if ticks_diff(ticks_ms(), b2.last_pressed()) >  300:
-                # long press
-                pass
-            else:
-                # short press
-                self.range += 1
-                if self.range > MAX_OUTPUT:
-                    self.range = MAX_OUTPUT
+    def x_scaled(self):
+        return (100.0 * (self.x - self.x_min))/self.x_range
 
-        # Freeze is triggered when din goes HIGH.
-        @din.handler
-        def dinTrigger():
-            # Pause
-            self.freeze = True
+    def y_scaled(self):
+        return (100.0 * (self.y - self.y_min))/self.y_range
 
-        @din.handler_falling
-        def dinTriggerEnd():
-            # Start agin
-            self.freeze = False
+    def z_scaled(self):
+        return (100.0 * (self.z - self.z_min))/self.z_range
+    
+    def __str__(self):
+        return (f"({self.x:2.2f},{self.y:2.2f},{self.z:2.2f})({self.s},{self.r},{self.b})")
 
-    def updateValues(self):
-        if not self.freeze:
-            self.a.step()
-        
-    def getSpeed(self):
-        # Set speed based on the knob.
-        # The range is piecewise linear from fully CCW to noon and noon to fully CW.
-        val = k1.read_position()
-        low = 1000 # CCW
-        mid = 100 # noon
-        high = 10 # CW
-        if val == 0:
-            self.period = low
-        elif val < 50:
-            self.period = low - ((low-mid) * (val/50))
-        else:
-            self.period = mid - ((mid-high) * (val-50)/50)
+    def step(self):
+        '''
+        Update the point.
+        '''
+        x_dot = self.s*(self.y - self.x)
+        y_dot = self.r*self.x - self.y - self.x*self.z
+        z_dot = self.x*self.y - self.b*self.z
+        self.x += x_dot * self.dt
+        self.y += y_dot * self.dt
+        self.z += z_dot * self.dt
 
-    def getThreshold(self):
-        val = k2.read_position()
-        self.threshold = (40 * val)/100
-        
+def main():
+    a = attractor()
+    a.estimateRanges()
+    print(a)
+    print(f"Min x:{a.x_min:8.2f} y:{a.y_min:8.2f} z:{a.z_min:8.2f}")
+    print(f"Max x:{a.x_max:8.2f} y:{a.y_max:8.2f} z:{a.z_max:8.2f}")
+    print(f"Ran x:{a.x_range:8.2f} y:{a.y_range:8.2f} z:{a.z_range:8.2f}")
 
-    def update(self):
-        # Change the values and output
-        self.updateValues()
-        cv1.voltage((self.range * self.a.x_scaled()) / 100)
-        cv2.voltage((self.range * self.a.y_scaled()) / 100)
-        cv3.voltage((self.range * self.a.z_scaled()) / 100)
-        # Calculate gates
-        # gate 1 fires if x is divisible by 2 when considered an int
-        self.gate4 = floor(self.a.x_scaled()) % 2 == 0
-        # gates 2 and 3 look at the differences between the outputs. 
-        self.gate5 = fabs(self.a.y_scaled() + self.a.z_scaled() - 2*self.a.x_scaled()) > self.threshold
-        self.gate6 = fabs(self.a.z_scaled() + self.a.x_scaled() - 2*self.a.y_scaled()) > self.threshold
-        if self.gate4:    
-            cv4.on()
-        else:
-            cv4.off()
-        if self.gate5:
-            cv5.on()
-        else:
-            cv5.off()
-        if self.gate6:
-            cv6.on()
-        else:
-            cv6.off()
-        
-        self.checkpoint = ticks_ms()
-        self.updateScreen()
-
-    def main(self):
-        while True:
-            self.getSpeed()
-            self.getThreshold()
-            if ticks_diff(ticks_ms(), self.checkpoint) > self.period:
-                self.update()
-
-    def initialiseMessage(self):
-        oled.fill(0)
-        oled.text('Initialising...',0,0,1)
-        oled.show()
-        
-    def updateScreen(self):
-        #oled.clear() - dont use this, it causes the screen to flicker!
-        oled.fill(0)
-        oled.text('1:' + str(int(self.a.x_scaled())),0,0,1)
-        oled.text('2:' + str(int(self.a.y_scaled())),0,8,1)
-        oled.text('3:' + str(int(self.a.z_scaled())),0,16,1)
-        oled.text('S:' + str(int(self.period)),40,0,1)
-        oled.text('T:' + str(int(self.threshold)),40,8,1)
-        oled.text('R:' + str(int(self.range)),40,16,1)
-        if self.gate4:
-            oled.text('4',100,0,1)
-        if self.gate5:
-            oled.text('5',100,8,1)
-        if self.gate6:
-            oled.text('6',100,16,1)
-        if self.freeze:
-            oled.text('FREEZE',0,24,1)
-            
-        oled.show()
-
-# Reset module display state.
-reset_state()
-loz = lorenz()
-loz.main()
+if __name__ == "__main__":
+    main()
